+++ packages/apidom-ls/src/config/openapi/link/lint/allowed-fields-3-1.ts
@@ -1,0 +1,24 @@
+import ApilintCodes from '../../../codes';
+import { LinterMeta } from '../../../../apidom-language-types';
+
+// eslint-disable-next-line @typescript-eslint/naming-convention
+const allowedFields3_1Lint: LinterMeta = {
+  code: ApilintCodes.NOT_ALLOWED_FIELDS,
+  source: 'apilint',
+  message: 'Object includes not allowed fields',
+  severity: 1,
+  linterFunction: 'allowedFields',
+  linterParams: [
+    ['operationRef', 'operationId', 'parameters', 'requestBody', 'description', 'server'],
+    'x-',
+  ],
+  marker: 'key',
+  conditions: [
+    {
+      function: 'missingField',
+      params: ['$ref'],
+    },
+  ],
+  targetSpecs: [{ namespace: 'openapi', version: '3.1.0' }],
+};
+export default allowedFields3_1Lint;